update hero component' git status
