+++ src/components/hero/hero.module.scss
@@ -1,6 +1,2 @@
 @import "../../styles/abstracts/variables";
 @import "../../styles/abstracts/mixins";
-
-.button {
-  @include link-full;
-}
