+++ src/views/home/home.js
@@ -1,10 +1,11 @@
 import { Header } from "../../components/header/header";
 import { Footer } from "../../components/footer/footer";
-import { HeroSection } from "../../components/hero/heroSection";
+import { Hero } from "../../components/hero/hero";
 import { DigitalBankingSection } from "../../components/digitalBanking/digitalBankingSection";
 import { StatsSection } from "../../components/statsSection/statsSection";
 import { SelectPackageSection } from "../../components/selectPackage/selectPackage";
 import { TextSection } from "../../components/textSection/textSection";
+import { PrivateBanker } from "../../components/privateBanker/privateBanker";
 export class Home {
   render() {
     const headerElement = document.getElementById("header");
@@ -16,8 +17,8 @@ export class Home {
       footer.setBurgerState(isActive);
     });
 
-    const heroBanner = new HeroSection();
-    heroBanner.render(mainElement);
+    const heroSection = new Hero();
+    heroSection.render(mainElement);
 
     const textSection = new TextSection();
     textSection.render(mainElement);
